more functionss

+++ ft_strlcpy.c
@@ -6,11 +6,13 @@
 /*   By: bcanals- <bcanals-@student.42barcel>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/02 14:30:19 by bcanals-          #+#    #+#             */
-/*   Updated: 2024/06/28 15:27:43 by bizcru           ###   ########.fr       */
+/*   Updated: 2024/07/01 20:51:06 by bizcru           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-long unsigned int	ft_strlcpy(char *dest, char *src, long unsigned int size)
+#include <stddef.h>
+
+size_t ft_strlcpy(char *dest, char *src, size_t size)
 {
 	unsigned int	i;
